# Speed up registerizeHarder for huge numbers of locals