Stage 2 of being fed up with long compile times. There's some cleanup (as promised in #3364), but the important commit for performance is "Pre-compute possible links and conflicts when considering a junction". The motivation for changes around here was profiling and observing that the highlighted two lines
```
          Block* block = blocks[b];
          Junction& jSucc = junctions[block->exit];
          for (auto otherName : jSucc.live) {      <<<<<<<<<<<<
            if (junc.live.has(otherName)) continue;    <<<<<<<<<<<
            if (block->lastKillLoc[otherName] < block->firstDeadLoc[name]) {

```
were really hot.
By doing preprocessing we can do map+vector iteration rather than map+map iteration. Given that the inner loop is run O(n^2) times, I'm guessing making it a vec loop is a Good Idea.
The actual motivating idea for the rearrangement is that, when checking for conflicts, we actually want to check for each variable against other variables. Having to loop over blocks and follow pointers to actually find the variables is an implementation detail of how the data is structured.

Before:
```
tests $ python runner.py asm3.test_sqlite
[...]
Ran 1 test in 73.380s
[...]
tests $ EMCC_DEBUG=1 emcc -O3 --llvm-lto 1 -s EMULATE_FUNCTION_POINTER_CASTS=1 -o out.js python/python.bc 2>&1 | grep 'js opts'
DEBUG    root: emcc step "js opts" took 6.09 seconds
```

After:
```
tests $ python runner.py asm3.test_sqlite
[...]
Ran 1 test in 60.924s
[...]
$ EMCC_DEBUG=1 emcc -O3 --llvm-lto 1 -s EMULATE_FUNCTION_POINTER_CASTS=1 -o out.js python/python.bc 2>&1 | grep 'js opts'
DEBUG    root: emcc step "js opts" took 4.48 seconds
```