@aidanhs 2015-04-17
test_other and the first 50 asm3 tests passed (+ sqlite, python). I'm going to run a full asm3 test.