@kripken 2015-04-17
Tests look good. I can confirm a 20% speedup (tested sqlite, openjpeg).