@kripken 2015-04-18
Ok, looks like everything is great here. I think there were some tiny tweaks from the comments, ready to merge when those are in.

Great stuff!