@aidanhs 2015-04-18
I have:
 - used insert to detect whether to reserve space in the set (which helped me remember why I hate C++...please check this code style for constructing objects etc is right)
 - used name rather than otherName
 - skipped entry junction for jWorkSet addition
 - put the (int) cast back in

I have not:
 - reserved space in jWorkSet - apparently can't do that in an ordered set. There are alternatives that I'll revisit if I see this loop getting hot.

Let me know if you have any more comments.