@aidanhs 2015-04-18
I'm annoyed github has swallowed a couple of helpful comments for future people looking back in history, so pasting here:

--------------

I wouldn't bother looking at the diff unless you're putting them side-by-side and considering them in isolation - this change actually alters the algorithm a bit.

To explain, let's consider 3 bocks (A, B and C) with their exit junctions (aJ, bJ, cJ). B and C are outblocks of aJ, i.e.:
```
   A
  / \
 B   C
```

To run through the two versions:

Previously:
```
For each live variable in aJ
  - mark conflicts against all live variables in aJ
  - for each successor block x
    - find xJ
    - for each live variable v in xJ
      - (optimisation: skip if it's live in aJ, it's already marked as a conflict)
      - if the variable lifetimes overlap in the block, mark them as conflicting
```

Note that we've got three nested loops - this is why the innermost stuff was so hot, because it gets executed a *lot*. Even the optimisation is touch and go - turn it into a slightly more expensive check (like checking if it's currently conflicting, rather than is just live in the current junction) and it slows down!
For example, imagine bJ and cJ have the same live variables - we're running the optimisation check twice as many times as necessary!

Now:
```
For each successor block x
  - find xJ
  - for each live variable v in xJ
    - add x as a possibly conflicting block for variable v
For each live variable v in aJ
  - remove v as a possibly conflicting variable
For each live variable v in aJ
  - mark conflicts against all live variables in aJ
  - for each potentially conflicting variable:
    - for each block the variable is present in
      - if the variable lifetimes overlap in the block, mark them as conflicting
```

We've effectively flipped the loop inside out, which has the advantage of being clearer - "for variable X, does Y conflict? let's check the blocks Y is present in".
On paper this kinda looks like more work, but I suspect the speedup comes from a) not having to follow pointers inside the doubly nested loop, b) iterating over a vector in the innermost loop but mainly c) moving the 'optimisation' checking whether variables are live up to a top-level loop rather than inside a thrice nested loop.

There are a few other tweaks I thought of while writing this, but they can wait.

-----------

The break is a 'free' optimisation as there's effectively no overhead in any situation.

I did consider adding a 'check if conflict exists' condition, but it had no noticeable effect I could detect and may cause slower running in some situations. We don't need it because we're inserting into a set - duplicates aren't an issue.

-----------

We must always enter this[do...while] loop in order to do a first pass consideration of any blocks added to bWorkSet above. The reason the condition moved to the end was exactly because of the exit junction case - even when we have an empty junction work set (i.e. it was just the exit junction), we need to look at the blocks going into the exit junction.