@kripken 2015-04-20
Note that they are hidden by default, but people can still see them, they just need to click on "expand outdated diff".