@aidanhs 2015-04-20
Hmm, the only "outdated diff" link I can see on this page is from rfk - I can't see your comments.
Did you click through to the individual commits and comment there?