@kripken 2015-04-20
Ah, yes, good point. I think github is smart about pull diff comments, but I commented on a commit - which is since the rebase not in this pull, I guess! You should get a separate email about it, but it doesn't show up here. Anyhow, all I said there was "makes sense, got it, thanks" or similar.